backend/requirements.txt
Django==3.2
python-socketio==5.2.1
eventlet==0.30.2
redis==3.5.3
djangorestframework==3.12.4
django-cors-headers==3.7.0
drf-yasg==1.20.0
djangorestframework-oauth==1.1.0
django-oauth-toolkit==1.5.0